django/language_learning_api/requirements.txt
# Language Learning API Requirements

Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
Pillow==10.0.1
psycopg2-binary==2.9.7

# Development dependencies
pytest==7.4.3
pytest-django==4.5.2
factory-boy==3.3.0
coverage==7.3.2

# Production dependencies
gunicorn==21.2.0
whitenoise==6.6.0